changes from PR comments
